classlib 2.1.2

+++ extlib/xgui/src/main/java/org/mini/gui/gscript/Int.java
@@ -27,7 +27,7 @@ public class Int extends DataType {
     }
 
     public long getVal() {
-        return (int) value;
+        return value;
     }
 
     public void setVal(long i) {
